# src/swagger.yaml
swagger: '2.0'
info:
  description: "This is the backen documentation for custommindz"
  version: 1.0.0
  title: "Customindz"
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
basePath: /dev
...
schemes:
  - https
  - http
paths:
...
  /monitor/:id/detection:
    get:
      tags:
       - "monitor"
      description: ""
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "monitor_id"
        in: "query"
        description: "monitor_id"
        required: true
        type: "string"
      - name: "orderBy"
        in: "query"
        description: "orderBy"
        type: "string"
      - name: "engine"
        in: "query"
        description: "engine"
        type: "string"
      - name: "start_timestamp"
        in: "query"
        description: "start_timestamp"
        type: "string"
      - name: "end_timestamp"
        in: "query"
        description: "end_timestamp"
        type: "string"
      - name: "alert"
        in: "query"
        description: "alert"
        type: "string"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/monitor"
        400:
          description: "err"
...
  /detection:
    get:
      tags:
       - "detection"
      description: ""
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "alert"
        in: "query"
        description: "alert"
        required: true
        type: "string"
      - name: "image_url"
        in: "query"
        description: "image_url"
        required: true
        type: "string"
      - name: "order"
        in: "query"
        description: "order"
        required: true
        type: "string"
      - name: "limit"
        in: "query"
        description: "limit"
        type: "string"
      - name: "offset"
        in: "query"
        description: "offset"
        type: "string"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/detection"
        400:
          description: "err"
  /detection/engine/:engineName:
    get:
      tags:
       - "detection"
      description: ""
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "monitor_id"
        in: "query"
        description: "monitor_id"
        required: true
        type: "string"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/detection"
        400:
          description: "err"
  /detection/:id:
    get:
      tags:
       - "detection"
      description: ""
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "query"
        description: "id"
        required: true
        type: "string"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/detection"
        400:
          description: "err"
    put:
      tags:
      - "detection"
      description: ""
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "id"
        required: true
        type: "string"
      - name: "body"
        in: "body"
        description: "Update"
        required: true
        schema:
          $ref: "#/definitions/monitor"
      responses:
        200:
          description: "Successfully Updated"
        400:
          description: "err"
    delete:
      tags:
      - "detection"
      summary: ""
      description: ""
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Successfully Deleted Monitor"
        400:
          description: "err"
  /detection/:id/vod:
    get:
      tags:
       - "detection"
      description: ""
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "query"
        description: "id"
        required: true
        type: "string"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/detection"
        400:
          description: "err"
definitions:
...
  detection:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      timestamp:
        type: "string"
      alert:
        type: "boolean"
      result:
        type: "string"
      monitor_id:
        type: "string"
      engine:
        type: "string"
      image_url:
        type: "string"
  monitor:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      company_code:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      connection_uri:
        type: "string"
      play_from_source:
        type: "boolean"
      recording:
        type: "boolean"
      engines:
        type: "string"
      graph:
        type: "string"
...